+++ pkgs/native_assets_cli/lib/locking.dart
@@ -1,8 +1,0 @@
-// Copyright (c) 2024, the Dart project authors.  Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
-/// A library that contains helpers for running hooks concurrently.
-library;
-
-export 'src/locking/locking.dart' show runUnderDirectoryLock;