+++ pkgs/native_assets_cli/lib/src/api/hook_config.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'dart:convert';
+import 'dart:io';
 
 import 'package:cli_config/cli_config.dart';
 import 'package:collection/collection.dart';
@@ -29,9 +30,29 @@ part '../model/hook_config.dart';
 /// This abstraction makes it easier to design APIs intended for both kinds of
 /// build hooks, building and linking.
 abstract class HookConfig {
-  /// The directory in which all output and intermediate artifacts should be
-  /// placed.
+  /// The directory in which output and intermediate artifacts that are unique
+  /// to this configuration can be placed.
+  ///
+  /// This directory is unique per hook and per configuration.
+  ///
+  /// The contents of this directory will not be modified by anything else than
+  /// the hook itself.
+  ///
+  /// The invoker of the the hook will ensure concurrent invocations wait on
+  /// each other.
   Uri get outputDirectory;
+
+  /// The directory in which shared output and intermediate artifacts can be
+  /// placed.
+  ///
+  /// This directory is unique per hook.
+  ///
+  /// The contents of this directory will not be modified by anything else than
+  /// the hook itself.
+  ///
+  /// The invoker of the the hook will ensure concurrent invocations wait on
+  /// each other.
+  Uri get outputDirectoryShared;
 
   /// The name of the package the assets are built for.
   String get packageName;