+++ pkgs/native_assets_cli/test/api/build_config_test.dart
@@ -14,6 +14,8 @@ void main() async {
   late Uri tempUri;
   late Uri outDirUri;
   late Uri outDir2Uri;
+  late Uri outputDirectoryShared;
+  late Uri outputDirectoryShared2;
   late String packageName;
   late Uri packageRootUri;
   late Uri fakeClang;
@@ -27,8 +29,12 @@ void main() async {
     outDirUri = tempUri.resolve('out1/');
     await Directory.fromUri(outDirUri).create();
     outDir2Uri = tempUri.resolve('out2/');
-    packageName = 'my_package';
     await Directory.fromUri(outDir2Uri).create();
+    outputDirectoryShared = tempUri.resolve('out_shared1/');
+    await Directory.fromUri(outputDirectoryShared).create();
+    outputDirectoryShared2 = tempUri.resolve('out_shared2/');
+    await Directory.fromUri(outputDirectoryShared2).create();
+    packageName = 'my_package';
     packageRootUri = tempUri.resolve('$packageName/');
     await Directory.fromUri(packageRootUri).create();
     fakeClang = tempUri.resolve('fake_clang');
@@ -50,6 +56,7 @@ void main() async {
   test('BuildConfig ==', () {
     final config1 = BuildConfig.build(
       outputDirectory: outDirUri,
+      outputDirectoryShared: outputDirectoryShared,
       packageName: packageName,
       packageRoot: tempUri,
       targetArchitecture: Architecture.arm64,
@@ -68,6 +75,7 @@ void main() async {
 
     final config2 = BuildConfig.build(
       outputDirectory: outDir2Uri,
+      outputDirectoryShared: outputDirectoryShared,
       packageName: packageName,
       packageRoot: tempUri,
       targetArchitecture: Architecture.arm64,
@@ -102,6 +110,7 @@ void main() async {
   test('BuildConfig fromConfig', () {
     final buildConfig2 = BuildConfig.build(
       outputDirectory: outDirUri,
+      outputDirectoryShared: outputDirectoryShared,
       packageName: packageName,
       packageRoot: packageRootUri,
       targetArchitecture: Architecture.arm64,
@@ -118,6 +127,7 @@ void main() async {
       'linking_enabled': false,
       'link_mode_preference': 'prefer-static',
       'out_dir': outDirUri.toFilePath(),
+      'out_dir_shared': outputDirectoryShared.toFilePath(),
       'package_name': packageName,
       'package_root': packageRootUri.toFilePath(),
       'target_android_ndk_api': 30,
@@ -133,6 +143,7 @@ void main() async {
   test('BuildConfig.dryRun', () {
     final buildConfig2 = BuildConfig.dryRun(
       outputDirectory: outDirUri,
+      outputDirectoryShared: outputDirectoryShared,
       packageName: packageName,
       packageRoot: packageRootUri,
       targetOS: OS.android,
@@ -146,6 +157,7 @@ void main() async {
       'linking_enabled': true,
       'link_mode_preference': 'prefer-static',
       'out_dir': outDirUri.toFilePath(),
+      'out_dir_shared': outputDirectoryShared.toFilePath(),
       'package_name': packageName,
       'package_root': packageRootUri.toFilePath(),
       'target_os': 'android',
@@ -159,6 +171,7 @@ void main() async {
   test('BuildConfig == dependency metadata', () {
     final buildConfig1 = BuildConfig.build(
       outputDirectory: outDirUri,
+      outputDirectoryShared: outputDirectoryShared,
       packageName: packageName,
       packageRoot: tempUri,
       targetArchitecture: Architecture.arm64,
@@ -180,6 +193,7 @@ void main() async {
 
     final buildConfig2 = BuildConfig.build(
       outputDirectory: outDirUri,
+      outputDirectoryShared: outputDirectoryShared,
       packageName: packageName,
       packageRoot: tempUri,
       targetArchitecture: Architecture.arm64,
@@ -208,6 +222,7 @@ void main() async {
   test('BuildConfig == hasLinkConfig', () {
     final buildConfig1 = BuildConfig.build(
       outputDirectory: outDirUri,
+      outputDirectoryShared: outputDirectoryShared,
       packageName: packageName,
       packageRoot: tempUri,
       targetArchitecture: Architecture.x64,
@@ -219,6 +234,7 @@ void main() async {
 
     final buildConfig2 = BuildConfig.build(
       outputDirectory: outDirUri,
+      outputDirectoryShared: outputDirectoryShared,
       packageName: packageName,
       packageRoot: tempUri,
       targetArchitecture: Architecture.x64,
@@ -236,6 +252,7 @@ void main() async {
   test('BuildConfig fromArgs', () async {
     final buildConfig = BuildConfig.build(
       outputDirectory: outDirUri,
+      outputDirectoryShared: outputDirectoryShared,
       packageName: packageName,
       packageRoot: tempUri,
       targetArchitecture: Architecture.arm64,