+++ pkgs/native_assets_cli/test/model/build_config_test.dart
@@ -16,6 +16,8 @@ void main() async {
   late Uri tempUri;
   late Uri outDirUri;
   late Uri outDir2Uri;
+  late Uri outputDirectoryShared;
+  late Uri outputDirectoryShared2;
   late Uri packageRootUri;
   late Uri fakeClang;
   late Uri fakeLd;
@@ -29,6 +31,10 @@ void main() async {
     await Directory.fromUri(outDirUri).create();
     outDir2Uri = tempUri.resolve('out2/');
     await Directory.fromUri(outDir2Uri).create();
+    outputDirectoryShared = tempUri.resolve('out_shared1/');
+    await Directory.fromUri(outputDirectoryShared).create();
+    outputDirectoryShared2 = tempUri.resolve('out_shared2/');
+    await Directory.fromUri(outputDirectoryShared2).create();
     packageRootUri = tempUri.resolve('$packageName/');
     await Directory.fromUri(packageRootUri).create();
     fakeClang = tempUri.resolve('fake_clang');
@@ -50,6 +56,7 @@ void main() async {
   test('BuildConfig ==', () {
     final config1 = BuildConfigImpl(
       outputDirectory: outDirUri,
+      outputDirectoryShared: outputDirectoryShared,
       packageName: packageName,
       packageRoot: tempUri,
       targetArchitecture: ArchitectureImpl.arm64,
@@ -67,6 +74,7 @@ void main() async {
 
     final config2 = BuildConfigImpl(
       outputDirectory: outDir2Uri,
+      outputDirectoryShared: outputDirectoryShared,
       packageName: packageName,
       packageRoot: tempUri,
       targetArchitecture: ArchitectureImpl.arm64,
@@ -99,6 +107,7 @@ void main() async {
   test('BuildConfig fromConfig', () {
     final buildConfig2 = BuildConfigImpl(
       outputDirectory: outDirUri,
+      outputDirectoryShared: outputDirectoryShared,
       packageName: packageName,
       packageRoot: packageRootUri,
       targetArchitecture: ArchitectureImpl.arm64,
@@ -115,6 +124,7 @@ void main() async {
       'linking_enabled': false,
       'link_mode_preference': 'prefer-static',
       'out_dir': outDirUri.toFilePath(),
+      'out_dir_shared': outputDirectoryShared.toFilePath(),
       'package_name': packageName,
       'package_root': packageRootUri.toFilePath(),
       'target_android_ndk_api': 30,
@@ -130,6 +140,7 @@ void main() async {
   test('BuildConfig.dryRun', () {
     final buildConfig2 = BuildConfigImpl.dryRun(
       outputDirectory: outDirUri,
+      outputDirectoryShared: outputDirectoryShared,
       packageName: packageName,
       packageRoot: packageRootUri,
       targetOS: OSImpl.android,
@@ -142,6 +153,7 @@ void main() async {
       'linking_enabled': false,
       'link_mode_preference': 'prefer-static',
       'out_dir': outDirUri.toFilePath(),
+      'out_dir_shared': outputDirectoryShared.toFilePath(),
       'package_name': packageName,
       'package_root': packageRootUri.toFilePath(),
       'target_os': 'android',
@@ -155,6 +167,7 @@ void main() async {
   test('BuildConfig toJson fromConfig', () {
     final buildConfig1 = BuildConfigImpl(
       outputDirectory: outDirUri,
+      outputDirectoryShared: outputDirectoryShared,
       packageName: packageName,
       packageRoot: packageRootUri,
       targetArchitecture: ArchitectureImpl.arm64,
@@ -177,6 +190,7 @@ void main() async {
   test('BuildConfig == dependency metadata', () {
     final buildConfig1 = BuildConfigImpl(
       outputDirectory: outDirUri,
+      outputDirectoryShared: outputDirectoryShared,
       packageName: packageName,
       packageRoot: tempUri,
       targetArchitecture: ArchitectureImpl.arm64,
@@ -198,6 +212,7 @@ void main() async {
 
     final buildConfig2 = BuildConfigImpl(
       outputDirectory: outDirUri,
+      outputDirectoryShared: outputDirectoryShared,
       packageName: packageName,
       packageRoot: tempUri,
       targetArchitecture: ArchitectureImpl.arm64,
@@ -225,6 +240,7 @@ void main() async {
     final outDir = outDirUri;
     final buildConfig1 = BuildConfigImpl(
       outputDirectory: outDir,
+      outputDirectoryShared: outputDirectoryShared,
       packageName: packageName,
       packageRoot: tempUri,
       targetArchitecture: ArchitectureImpl.arm64,
@@ -263,6 +279,7 @@ void main() async {
       'linking_enabled': false,
       'link_mode_preference': 'prefer-static',
       'out_dir': outDir.toFilePath(),
+      'out_dir_shared': outputDirectoryShared.toFilePath(),
       'package_name': packageName,
       'package_root': tempUri.toFilePath(),
       'supported_asset_types': [NativeCodeAsset.type],
@@ -304,6 +321,7 @@ target_os: ios
 version: 1.0.0''';
     final buildConfig1 = BuildConfigImpl(
       outputDirectory: outDir,
+      outputDirectoryShared: outDir.resolve('../out_shared/'),
       packageName: packageName,
       packageRoot: tempUri,
       targetArchitecture: ArchitectureImpl.arm64,
@@ -367,6 +385,7 @@ version: 1.0.0''';
       () => BuildConfigImpl.fromJson({
         'version': HookConfigImpl.latestVersion.toString(),
         'out_dir': outDirUri.toFilePath(),
+        'out_dir_shared': outputDirectoryShared.toFilePath(),
         'package_name': packageName,
         'package_root': packageRootUri.toFilePath(),
         'target_architecture': 'arm64',
@@ -391,6 +410,7 @@ version: 1.0.0''';
     expect(
       () => BuildConfigImpl.fromJson({
         'out_dir': outDirUri.toFilePath(),
+        'out_dir_shared': outputDirectoryShared.toFilePath(),
         'version': HookConfigImpl.latestVersion.toString(),
         'package_name': packageName,
         'package_root': packageRootUri.toFilePath(),
@@ -412,6 +432,7 @@ version: 1.0.0''';
   test('BuildConfig toString', () {
     final config = BuildConfigImpl(
       outputDirectory: outDirUri,
+      outputDirectoryShared: outputDirectoryShared,
       packageName: packageName,
       packageRoot: tempUri,
       targetArchitecture: ArchitectureImpl.arm64,
@@ -431,6 +452,7 @@ version: 1.0.0''';
   test('BuildConfig fromArgs', () async {
     final buildConfig = BuildConfigImpl(
       outputDirectory: outDirUri,
+      outputDirectoryShared: outputDirectoryShared,
       packageName: packageName,
       packageRoot: tempUri,
       targetArchitecture: ArchitectureImpl.arm64,
@@ -454,6 +476,7 @@ version: 1.0.0''';
   test('envScript', () {
     final buildConfig1 = BuildConfigImpl(
       outputDirectory: outDirUri,
+      outputDirectoryShared: outputDirectoryShared,
       packageName: packageName,
       packageRoot: packageRootUri,
       targetArchitecture: ArchitectureImpl.x64,
@@ -479,6 +502,7 @@ version: 1.0.0''';
       final config = {
         'link_mode_preference': 'prefer-static',
         'out_dir': outDir.toFilePath(),
+        'out_dir_shared': outputDirectoryShared.toFilePath(),
         'package_root': tempUri.toFilePath(),
         'target_os': 'linux',
         'version': version,
@@ -502,6 +526,7 @@ version: 1.0.0''';
     final config = {
       'link_mode_preference': 'prefer-static',
       'out_dir': outDir.toFilePath(),
+      'out_dir_shared': outputDirectoryShared.toFilePath(),
       'package_name': packageName,
       'package_root': tempUri.toFilePath(),
       'target_os': 'windows',
@@ -522,6 +547,7 @@ version: 1.0.0''';
     final config = {
       'link_mode_preference': 'prefer-static',
       'out_dir': outDir.toFilePath(),
+      'out_dir_shared': outputDirectoryShared.toFilePath(),
       'package_name': packageName,
       'package_root': tempUri.toFilePath(),
       'target_os': 'windows',
@@ -546,6 +572,7 @@ version: 1.0.0''';
       'linking_enabled': true,
       'link_mode_preference': 'prefer-static',
       'out_dir': outDir.toFilePath(),
+      'out_dir_shared': outputDirectoryShared.toFilePath(),
       'package_name': packageName,
       'package_root': tempUri.toFilePath(),
       'target_os': 'android',
@@ -567,6 +594,7 @@ version: 1.0.0''';
       'linking_enabled': false,
       'link_mode_preference': 'prefer-static',
       'out_dir': outDir.toFilePath(),
+      'out_dir_shared': outputDirectoryShared.toFilePath(),
       'package_name': packageName,
       'package_root': tempUri.toFilePath(),
       'target_os': 'windows',
@@ -579,6 +607,7 @@ version: 1.0.0''';
   test('BuildConfig dry_run toString', () {
     final buildConfig = BuildConfigImpl.dryRun(
       packageName: packageName,
+      outputDirectoryShared: outputDirectoryShared,
       outputDirectory: outDirUri,
       packageRoot: tempUri,
       targetOS: OSImpl.windows,
@@ -595,6 +624,7 @@ version: 1.0.0''';
       'dry_run': false,
       'link_mode_preference': 'prefer-static',
       'out_dir': outDirUri.toFilePath(),
+      'out_dir_shared': outputDirectoryShared.toFilePath(),
       'package_name': packageName,
       'package_root': packageRootUri.toFilePath(),
       'target_android_ndk_api': 30,